[25] Use typescript (#26)
This commit moves javascript files into a `src` directory.
All .js files in src (with the exception of models) have been converted to .ts files.
models have not been converted, as they are planned to be rewritten in another issue.

This commit updates config files, and scripts to work with typescript. tsconfig.json is created.