[4] Use config files (#43)
I created botConfig.json and defaultBotConfig.json. defaultBotConfig.json contains the "out of the box" settings for my bot. A setting defined in botConfig.json overrides default settings. If a setting is not defined in botConfig.json, the setting defined in defaultBotConfig.json is used.

CommandRouter
I refactored CommandRouter to be a class, and use config files. I injected config files into the constructor; This helped with testing the logic.